refactor: create widgets folder (asset, cost)
