+++ apps/web/src/services/dashboards/widgets/_configs/widget-list-config.ts
@@ -3,25 +3,25 @@ import baseCountOfFindings from '@/services/dashboards/widgets/_base/base-count-
 import basePie from '@/services/dashboards/widgets/_base/base-pie/widget-config';
 import baseTrend from '@/services/dashboards/widgets/_base/base-trend/widget-config';
 // console widgets
-import awsCloudFrontCost from '@/services/dashboards/widgets/aws-cloud-front-cost/widget-config';
-import awsDataTransferByRegion from '@/services/dashboards/widgets/aws-data-transfer-by-region/widget-config';
-import awsDataTransferCostTrend from '@/services/dashboards/widgets/aws-data-transfer-cost-trend/widget-config';
-import budgetStatus from '@/services/dashboards/widgets/budget-status/widget-config';
-import budgetUsageByTarget from '@/services/dashboards/widgets/budget-usage-by-target/widget-config';
-import budgetUsageSummary from '@/services/dashboards/widgets/budget-usage-summary/widget-config';
-import complianceCheckStatus from '@/services/dashboards/widgets/compliance-check-status/widget-config';
-import costByRegion from '@/services/dashboards/widgets/cost-by-region/widget-config';
-import costDonut from '@/services/dashboards/widgets/cost-donut/widget-config';
-import costMap from '@/services/dashboards/widgets/cost-map/widget-config';
-import costPie from '@/services/dashboards/widgets/cost-pie/widget-config';
-import costTrendStacked from '@/services/dashboards/widgets/cost-trend-stacked/widget-config';
-import costTrend from '@/services/dashboards/widgets/cost-trend/widget-config';
-import countOfFailFindings from '@/services/dashboards/widgets/count-of-fail-findings/widget-config';
-import countOfPassAndFailFindings from '@/services/dashboards/widgets/count-of-pass-and-fail-findings/widget-config';
-import monthlyCost from '@/services/dashboards/widgets/monthly-cost/widget-config';
-import severityStatusByService from '@/services/dashboards/widgets/severity-status-by-service/widget-config';
-import totalFailureAndSeverity from '@/services/dashboards/widgets/total-failure-and-severity/widget-config';
-import trendOfPassAndFailFindings from '@/services/dashboards/widgets/trend-of-pass-and-fail-findings/widget-config';
+import complianceCheckStatus from '@/services/dashboards/widgets/asset-widgets/compliance-check-status/widget-config';
+import countOfFailFindings from '@/services/dashboards/widgets/asset-widgets/count-of-fail-findings/widget-config';
+import countOfPassAndFailFindings from '@/services/dashboards/widgets/asset-widgets/count-of-pass-and-fail-findings/widget-config';
+import severityStatusByService from '@/services/dashboards/widgets/asset-widgets/severity-status-by-service/widget-config';
+import totalFailureAndSeverity from '@/services/dashboards/widgets/asset-widgets/total-failure-and-severity/widget-config';
+import trendOfPassAndFailFindings from '@/services/dashboards/widgets/asset-widgets/trend-of-pass-and-fail-findings/widget-config';
+import awsCloudFrontCost from '@/services/dashboards/widgets/cost-widgets/aws-cloud-front-cost/widget-config';
+import awsDataTransferByRegion from '@/services/dashboards/widgets/cost-widgets/aws-data-transfer-by-region/widget-config';
+import awsDataTransferCostTrend from '@/services/dashboards/widgets/cost-widgets/aws-data-transfer-cost-trend/widget-config';
+import budgetStatus from '@/services/dashboards/widgets/cost-widgets/budget-status/widget-config';
+import budgetUsageByTarget from '@/services/dashboards/widgets/cost-widgets/budget-usage-by-target/widget-config';
+import budgetUsageSummary from '@/services/dashboards/widgets/cost-widgets/budget-usage-summary/widget-config';
+import costByRegion from '@/services/dashboards/widgets/cost-widgets/cost-by-region/widget-config';
+import costDonut from '@/services/dashboards/widgets/cost-widgets/cost-donut/widget-config';
+import costMap from '@/services/dashboards/widgets/cost-widgets/cost-map/widget-config';
+import costPie from '@/services/dashboards/widgets/cost-widgets/cost-pie/widget-config';
+import costTrendStacked from '@/services/dashboards/widgets/cost-widgets/cost-trend-stacked/widget-config';
+import costTrend from '@/services/dashboards/widgets/cost-widgets/cost-trend/widget-config';
+import monthlyCost from '@/services/dashboards/widgets/cost-widgets/monthly-cost/widget-config';
 // eslint-disable-next-line import/order
 import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
