+++ apps/web/src/services/dashboards/widgets/asset-widgets/compliance-check-status/ComplianceCheckStatusWidget.vue
