+++ apps/web/src/services/dashboards/widgets/cost-widgets/aws-cloud-front-cost/widget-config.ts
@@ -10,7 +10,7 @@ import {
 const awsCloudFrontCostWidgetConfig: WidgetConfig = {
     widget_config_id: 'awsCloudFrontCost',
     widget_component: () => ({
-        component: import('@/services/dashboards/widgets/aws-cloud-front-cost/AWSCloudFrontCostWidget.vue'),
+        component: import('@/services/dashboards/widgets/cost-widgets/aws-cloud-front-cost/AWSCloudFrontCostWidget.vue'),
     }),
     title: 'AWS CloudFront Cost',
     labels: ['Cost'],
