+++ apps/web/src/services/dashboards/widgets/cost-widgets/budget-status/widget-config.ts
@@ -1,29 +1,31 @@
+import { GRANULARITY } from '@/services/dashboards/config';
 import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
-import { GRANULARITY } from '@/services/dashboards/widgets/_configs/config';
+import { CHART_TYPE } from '@/services/dashboards/widgets/_configs/config';
 import {
     getWidgetFilterOptionsSchema,
     getWidgetFilterSchemaPropertyNames,
 } from '@/services/dashboards/widgets/_helpers/widget-schema-helper';
 
-const budgetUsageSummaryConfig: WidgetConfig = {
-    widget_config_id: 'budgetUsageSummary',
+const budgetStatusWidgetConfig: WidgetConfig = {
+    widget_config_id: 'budgetStatus',
     widget_component: () => ({
-        component: import('@/services/dashboards/widgets/budget-usage-summary/BudgetUsageSummaryWidget.vue'),
+        component: import('@/services/dashboards/widgets/cost-widgets/budget-status/BudgetStatus.vue'),
     }),
-    title: 'Budget Usage Summary',
+    title: 'Budget Status',
     labels: ['Cost'],
     description: {
-        translation_id: 'DASHBOARDS.WIDGET.BUDGET_USAGE_SUMMARY.DESC',
-        preview_image: 'widget-img_budgetUsageSummary--thumbnail.png',
+        translation_id: 'DASHBOARDS.WIDGET.BUDGET_STATUS.DESC',
+        preview_image: 'widget-img_budgetStatus--thumbnail.png',
     },
     scopes: ['PROJECT', 'WORKSPACE'],
     theme: {
-        inherit: true,
-        inherit_count: 1,
+        inherit: false,
     },
-    sizes: ['sm', 'full'],
+    sizes: ['sm'],
     options: {
+        chart_type: CHART_TYPE.WAFFLE,
         granularity: GRANULARITY.ACCUMULATED,
+        cost_group_by: 'budget_id',
     },
     options_schema: {
         default_properties: getWidgetFilterSchemaPropertyNames('provider', 'project', 'region', 'cost_product'),
@@ -35,4 +37,4 @@ const budgetUsageSummaryConfig: WidgetConfig = {
     },
 };
 
-export default budgetUsageSummaryConfig;
+export default budgetStatusWidgetConfig;
