+++ apps/web/src/services/dashboards/widgets/cost-widgets/cost-map/CostMapWidget.vue
