+++ apps/web/src/services/dashboards/widgets/cost-widgets/cost-trend-stacked/widget-config.ts
