+++ apps/web/src/services/dashboards/widgets/cost-widgets/cost-trend/widget-config.ts
