+++ apps/web/src/services/dashboards/widgets/cost-widgets/monthly-cost/widget-config.ts
@@ -1,38 +1,35 @@
 import type { WidgetConfig } from '@/services/dashboards/widgets/_configs/config';
-import { GRANULARITY, COST_GROUP_BY } from '@/services/dashboards/widgets/_configs/config';
+import { GRANULARITY } from '@/services/dashboards/widgets/_configs/config';
 import {
-    getWidgetFilterOptionsSchema, getWidgetFilterSchemaPropertyNames,
-    getWidgetOptionsSchema,
+    getWidgetFilterOptionsSchema,
+    getWidgetFilterSchemaPropertyNames,
 } from '@/services/dashboards/widgets/_helpers/widget-schema-helper';
 
-const costMapWidgetConfig: WidgetConfig = {
-    widget_config_id: 'costMap',
+const monthlyCostWidgetConfig: WidgetConfig = {
+    widget_config_id: 'monthlyCost',
     widget_component: () => ({
-        component: import('@/services/dashboards/widgets/cost-map/CostMapWidget.vue'),
+        component: import('@/services/dashboards/widgets/cost-widgets/monthly-cost/MonthlyCostWidget.vue'),
     }),
-    title: 'Cost Map',
+    title: 'Monthly Cost',
     labels: ['Cost'],
     description: {
-        translation_id: 'DASHBOARDS.WIDGET.COST_MAP.DESC',
-        preview_image: 'widget-img_costMap--thumbnail.png',
+        translation_id: 'DASHBOARDS.WIDGET.MONTHLY_COST.DESC',
+        preview_image: 'widget-img_monthlyCost--thumbnail.png',
     },
-    scopes: ['PROJECT', 'WORKSPACE'],
+    scopes: ['DOMAIN', 'PROJECT', 'WORKSPACE'],
     theme: {
         inherit: true,
         inherit_count: 1,
     },
-    sizes: ['md', 'full'],
+    sizes: ['sm'],
     options: {
-        cost_group_by: COST_GROUP_BY.PROJECT,
-        granularity: GRANULARITY.ACCUMULATED,
+        granularity: GRANULARITY.MONTHLY,
     },
     options_schema: {
-        default_properties: ['cost_group_by', ...getWidgetFilterSchemaPropertyNames('provider', 'project', 'service_account', 'region', 'cost_product', 'cost_account')],
-        fixed_properties: ['cost_group_by'],
+        default_properties: getWidgetFilterSchemaPropertyNames('provider', 'project', 'service_account', 'region', 'cost_product', 'cost_account'),
         schema: {
             type: 'object',
             properties: {
-                ...getWidgetOptionsSchema('cost_group_by'),
                 ...getWidgetFilterOptionsSchema(
                     'provider',
                     'project',
@@ -46,7 +43,7 @@ const costMapWidgetConfig: WidgetConfig = {
                     'cost_account',
                 ),
             },
-            order: ['cost_group_by', ...getWidgetFilterSchemaPropertyNames(
+            order: getWidgetFilterSchemaPropertyNames(
                 'provider',
                 'project',
                 'service_account',
@@ -57,9 +54,9 @@ const costMapWidgetConfig: WidgetConfig = {
                 'region',
                 'cost_type',
                 'cost_account',
-            )],
+            ),
         },
     },
 };
 
-export default costMapWidgetConfig;
+export default monthlyCostWidgetConfig;
